requirements.txt
Django~=6.0.5
djangorestframework~=3.17.1
django-filter~=25.2
drf-spectacular~=0.29.0
djangorestframework_simplejwt~=5.5.1
python-decouple~=3.8
psycopg2
gunicorn